feat,ref,test(game): Add new columns to entities. Wrap controller's response in json. This commit added new column to entities: money_per_pull. This change allows a game to calculate the amount that the user has to spend per pull.
This commit added new column to entities: Banner Pity System, Game Character.
This change allows a banner to specify which type of character is allowed for
each rarity: CHARACTER, WEAPON, or ALL.

This commit update the util function transformGame() and transformBanners()
to incorporate the data change and new features from them.

This commit wrap response from GamesController in data.

This commit rewrite the test cases to match the changes.